Merge remote-tracking branch 'origin/master'
# Conflicts:
#	src/screens/MainMenu.java
Switched each screen to extending JukeScreen, which implements JukeScreenIF and makes it easier to deal with the dynamic typing and static types of the main pane inside the ScreenController. Also add some functionality to buttons. Lastly, i had to comment out some errors i had in the AdminAddSong screen (I think its specific to intellij, so feel free to uncomment it)